+++ pyproject.toml
@@ -15,6 +15,7 @@ build-backend = "poetry.core.masonry.api"
 
 boto3 = "^1.18.3"
 botocore = "^1.29.84"
+coverage = "^7.2.2"
 pytest = "^"
 unittest = "^"
 sys = "^"
